refactor: move types into own file in api crate
